implements checksum.
